fix: useLoader shows previous error messages if data is already loaded
The useLoader hook has a bug which causes error messages to persist when switching to a new item, IF the new item has already been loaded.

This is because of the `if (item == null)` check that skips everything if the item is already loaded.

This commit assumes that if the item is already loaded, the error states should be cleared.