feat: Stats logging of the number of queries with INITIALIZED status
This adds a new metric in the stats_logger system, `query.initialized`, which tracks the number of QueryExecutions with a status of `INITIALIZED`.  This indicates whether or not there's a queue on the Querybook side to submit queries to the Trino cluster.

This metric can be monitored to alert when there might be an issue with the Querybook workers not functioning properly.