fix: Catch exceptions during delete_schema/table_not_in_metastore
At the beginning of the metastore sync, these functions are called to remove schemas and tables that are no longer found in the Hive Metastore.  Previously, they were rewritten to do a double-check against the HMS to verify that the schema/table truly no longer exists.

This potentially adds a lot of additional HMS calls, so any issue with the HMS server potentially can break the process.

Since releasing those changes, we've seen intermittent failures due to `Internal error processing get_database`.

To address this, this PR adds a `try`/`except` around each schema/table, so that a one-off failure won't stop the entire task.  Worse case is that it will skip deleting a schema or table, but this can be self-corrected during the next sync.