fix: Boost JSON column detector priority to ensure it matches
Since the JSON detector is limited to `object` types (arrays, objects), it doesn't match things like numbers/booleans/strings.  This change ensures that if the content in the column looks like a JSON array or object, it will be rendered with the JSON transform.

This is a better change than updating the `id` detection, because if a JSON column was called `id` we would still want to use the JSON transform.